+++ package.json
@@ -3,9 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.27",
+    "@types/react": "^18.0.7",
+    "@types/react-dom": "^18.0.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-scripts": "3.3.1"
+    "react-scripts": "3.3.1",
+    "typescript": "^4.6.3"
   },
   "scripts": {
     "start": "react-scripts start",
