+++ src/index.js
@@ -3,14 +3,15 @@ import ReactDOM from "react-dom";
 
 import "./index.css";
 
-import GithubIssueComments from "./GithubIssueComments/GithubIssueComments";
+import GithubIssueComments from "./GithubIssueComments";
 
 ReactDOM.render(
   <GithubIssueComments
-    issueUri="sveltejs/svelte/issues/2546"
-    useShowCommentsButton={true}
+    issueUri="Pjaerr/Svelte-Electron-Desktop-App/issues/1"
+    useShowCommentsButton={false}
     commentsPerPage={5}
     allowRefreshingComments={true}
+    direction="asc"
   />,
   document.getElementById("root")
 );
